LogRequestConsole - Added LOG request type, will be uploaded to server via /api/log/{level}
OutboundNetworkManager - comment update

NetworkManager - added log to api test dropdown
