+++ frontend/src/API/OutboundNetworkHandler.ts
@@ -13,7 +13,6 @@ export const BACKEND_BASE: string = "http://localhost:3000/api";
  *
  */
 export async function sendPostRequest(endpoint: string, body?: unknown) {
-    // Scanner
     // Backend
     const response = await fetch(BACKEND_BASE + endpoint, {
         method: "POST",
@@ -37,7 +36,6 @@ export async function sendPostRequest(endpoint: string, body?: unknown) {
  *
  */
 export async function sendGetRequest(endpoint: string) {
-    // Scanner
     // Backend
     const response = await fetch(BACKEND_BASE + endpoint, {
         method: "GET",
